Added minimum build option
Necessary for building a project with bare minimum options. Having this
option alleviates the need to type the following:

    rubysmith --build demo --no-bundler-audit \
                           --no-console \
                           --no-documentation \
                           --no-git \
                           --no-git-lint \
                           --no-guard \
                           --no-reek \
                           --no-pry \
                           --no-rubocop \
                           --no-rspec \
                           --no-simple_cov \
                           --no-setup

Now one can type the following to achieve the above:

    rubysmith --build demo --min

Having a `--min` build option provides an experience that is nearly the
equivalent of a Bundler Inline script but without having to check gem
dependencies each time the code is run.

Use of *Style/MethodCallWithArgsParentheses* is temporarily disabled in
the shell due to a bug with pattern matching.