Added RubyCritic
Provides additional metrics and code quality analysis on top of what is
produced by SimpleCov. This is useful when needing a high level visual
analysis of a project.

Included in these changes is both direct support for this project to
benefit from RubyCritic and a new `--ruby_critic` build option when
building new projects using Rubysmith.