Refactored loading of configuration
Necessary to define a distinct namespace for configuration management
along with a specialized loader. The introduction of this namespace
will allow us to add more objects, in the future, which will define a
better structure for the data being loaded from the configuration.

This also mimics the work recently applied to the Tocer gem.