Added CLI configuration content project path
Necessary to obtain the relative path of project, especially when the
project has a dash in the name. Examples:

    # Name     | Path
    # -------------------
    # git      | git
    # git_lint | git_lint
    # git-lint | git/lint

Notice, in the last example, a nested directory structure is used. This
behavior is new and matches Gemsmith behavior when creating gems.

Additional specs were added, for clarification, to all methods effected
by single, underscored, and/or dashed project names.