Updated Reek and Rubocop configuration to ignore CLI build parser
Both Reek and Rubocop have flagged this object as having too many
methods but there isn't a strong reason to break this object up any
further. Having all build options located in a single object improves
the maintenance and look up of CLI options as a single source of truth.