Added Tocer to Gemfile skeleton when Markdown format is enabled
Now that Markdown files -- at least for the README -- have Table of
Contents generated for them when building a project skeleton, we can
include this gem within the project too. By having this gem included,
it will be easier for recently built projects to update their Table of
Contents accordingly.