Updated builders to answer configuration
Necessary to answer context with useful information than `nil`. The
configuration is what drives these builders so is handy to have these
builder always answer the same object. This also captures behavior with
minimum and maximum configurations as, previously, the builders were
not consistent in what was returned when maximum options were given.