Updated configuration content to support multiple template roots
Necessary to allow multiple template roots to be defined so there can
be any number of fallback paths. The first element in the array will
be checked first. If the path doesn't exist, the next path will be
checked and so forth. This works a lot like the Runcom gem which is
built on top of XDG where any local path will take precidence over any
global path.

Any new path added will be the first path to be checked.