Fixed pending CI shell specs
Necessary to ensure these specs run locally as well as on CI. The
caveat is that Bundler won't build the `Gemfile.lock` when on CI for
an unknown reason.

The file filter was updated to remove more Git files since we only care
the repository is created properly. Looking for the `HEAD` file is
enough to verify this is correct.

With this change all remaining pending specs are removed.