Added Circle CI SimpleCov artifacts to template
Necessary to ensure artifacts are created for Circle CI configurations
which also have SimpleCov enabled. This matches the changes applied to
this project for improved consistency.

There is a slight inconsistency with this change in that the ERB
conditional is *not indented* like other templates. This was done for
simplicity but should be made consistent in the future.