Updated loader enhancers injection to use container
Now that we have a container of enhancers we'll inject them for processing but instead of itereating through an array, we iterate in a hash-like manner. These are not being injected via Infusable because knowning the key associated with the enhancer will be important for upstream dependnecies, like Gemsmith, to obtain and manipulate further.