Updated configuration loader to supply template root for content
Completes the work started in the previous commit where the `TemplateRoot` enhancer is now used to supply default information. This also means that the `Content` class no longer need to supply a default but use an empty array instead. The `Content#initialize` method was updated to leverage Ruby 3.2.0 syntax for keyword arguments along with safe defaults.