Updated configuration to use strings instead of symbols
This is in preparation for adding Etcher configuration load support where the configuration will be safely loaded which does not allow for symbols. This also includes updates to the Pragmater and Tocer extensions since both use the `patterns` key instead of `includes` now. Optional keys have been deleted since they will be verified by the contract (soon to be implemented).