Added configuration target root transformer
Necessary to extract behavior to this transformer so the target root can default to the current directory when the key is missing (default behavior). The target root can be customized/overwritten, if desired, but is not recommended and is only provided for advanced use cases.

Having this transformer in place will allow the model initializer to be simplified further.