Fixed Git-based transformers to be lenient with missing values
Necessary to address issues with the introduction of the configuration contract where all of these Git-based keys are optional but must be filled if present. The problem was when asking Git for this information since the default was to answer an empty string when not present but that would violate the contract. These changes allow each transformer to make multiple attempts to acquire valid values and, if all else fails, the original content will be answered back as a success instead of failure.