Added builder check
Provides the ability to check if a build path exists and abort the entire build process if so. This will soon be used by the builders to verify if a project already exists to prevent people from accidentally generating the same project in the current directory when it already exists.

The RuboCop check has been disabled, for now, because it's easier to keep all logic within the same class but this will need more consideration in the future.