Fixed CLI actions to use nil instead of default values
Due to actions being nearly a one-to-one mapping to `OptionParser#on` methods, it can seem like default values would be accepted but option parser ignores them. This can be confusing since, originally, these actions made it look like that was possible. In order to reduce confusion, I've opted to make all of these `nil` by default.