Fixed console script project requirement
Necessary to ensure the console loads properly for a newly built project. The use of `Bundler.root` to obtain the absolute path is necessary to ensure the project namespace is loaded. This behavior isn't necessary for gem projects since the path can be relative. The behavior between this project and the Gemsmith project differ in this regard.

Milestone: patch