Fixed gem loader to find by tag and cache instance
Necessary to fix situations where you want to acquire the loader in other projects. In this case, the loader is acquired by unique tag and then cached. Additionally specs have been added for proper coverage.

The template specs deal with whitespace as best as possible but all of this will need to be smoothed out more in the future. RuboCop will catch and auto-correct all of this when the project skeleton is built but is not ideal.

Milestone: patch