Removed extension configurations
Originally this feature was meant to allow customization of each extension but, in practice, this hasn't proven to be useful since each individual extension can be customized independently, if desired, via the extension's global or local configuration. Project maintenance is further reduced by this change since there are compatiblity issues now.

Milestone: major