Fixed Pragmater extension configuration
This was an unfortunate oversight where removing the original embedded configuration associated with this gem fell back to Pragmater's default configuration. We need to have a specialized configuration for this gem instead which this now addresses properly. This includes ensuring Pragamater is run within the newly generated project skeleton root directory.

Milestone: patch