Updated builder path/run methods to log info
Necessary to provide more insight into what files and folders are being created. This is a delicate balance as to what is too much information and what isn't. In this case, info logging is applied to methods which create files/folders or run commands. Methods that manipulate files/folders are left as debug messages. Hopefully this is enough information to be informative without being too verbose.

Milestone: minor