Updated abstract builder to inject logger
Necessary to pass the logger onto the builder so upstream projects/gems can provide their own custom logger as desired. In this case, I'm using a lambda as an adapter to pass dependencies so upstream objects can use the same interface. This will also allow us to remove the `Builder.call` method in the next major release.

Milestone: minor