Fixed Bundler extension to account for unstable internet connection
Necessary in situations where there is no internet or a connection that is unstable which causes Bundler to fail. In these situations, we can display the error but complete the building of the skeleton project because once internet is restored, you can manually run `bundle install` to resolve.

Milestone: patch